requirements.txt
--allow-external mysql-connector-python
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
gunicorn==19.1.1
itsdangerous==0.24
mysql-connector-python==2.0.3
psycopg2==2.5.4
requests==2.4.3
wsgiref==0.1.2
